requirements.txt
fastapi
pydantic
llama-cpp-python
yt-dlp
uvicorn
requests
opencv-python
python-dotenv
ffmpeg-python
openai-whisper